requirements.txt
bcrypt==4.0.1
bottle==0.12.23
gitpython==3.1.27
PyJWT==2.6.0
python-magic-bin==0.4.14
python-magic==0.4.27
requests==2.28.2
Pillow==9.5.0